buffered => parse, rename inner fns and remove low-value methods
